requirements.txt
atomicwrites==1.2.1
attrs==18.2.0
cachetools==3.0.0
certifi==2018.11.29
chardet==3.0.4
Click==7.0
coverage==4.5.2
google-api-core==1.7.0
google-auth==1.6.2
google-cloud-vision==0.35.2
googleapis-common-protos==1.5.5
grpcio==1.17.1
idna==2.8
more-itertools==5.0.0
Naked==0.1.31
numpy==1.15.4
opencv-python==3.4.5.20
pbr==5.1.1
pluggy==0.8.0
protobuf==3.6.1
py==1.7.0
pyasn1==0.4.5
pyasn1-modules==0.2.3
pycryptodome==3.7.2
pytest==4.1.0
pytest-cov==2.6.1
python-coveralls==2.9.1
pytz==2018.9
PyYAML==3.13
requests==2.21.0
rsa==4.0
shellescape==3.4.1
six==1.12.0
urllib3==1.24.1
